% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{japan_contacts}
\alias{japan_contacts}
\title{Age-specific mixing patterns for Japan}
\format{A tibble with twelve columns and twelve rows,
  representing the contact matrix}
\source{
{https://doi.org/10.1136/jech-2015-205777}
}
\usage{
japan_contacts
}
\description{
Population-weighted number of contacts between age groups
The data are described in Ibuka Y, et al. (2016) Social contacts, 
vaccination decisions and influenza in Japan. 
J Epidemiol Community Health 70:162-167.
}
\keyword{datasets}
