% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.r
\name{scenarios}
\alias{scenarios}
\title{Explore vaccination scenarios}
\usage{
scenarios(contacts, ...)
}
\arguments{
\item{contacts}{a contact matrix}

\item{...}{any further parameters to pass to \code{\link{sk_projected_seroprevalence}}}
}
\value{
wide table of seroprevalence by age group, original (2014) and shifted (2018)
}
\description{
Explore vaccination scenarios
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
