% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{sk_mcv_coverage}
\alias{sk_mcv_coverage}
\title{Measles vaccination coverage data for South Korea}
\format{A tibble with three columns: vaccine (MCV1 or MCV2 for
  first/second dose, year and uptake.}
\source{
\url{https://www.who.int/immunization/monitoring_surveillance/data/en/}
}
\usage{
sk_mcv_coverage
}
\description{
A dataset containing measles vaccination coverage in South Korea
as reported by the World Health Organization
}
\keyword{datasets}
