% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{sk_population}
\alias{sk_population}
\title{Population by age for South Korea}
\format{A tibble with three columns: lower.age.limit (the lower bound
  in years of the age group), year and population}
\usage{
sk_population
}
\description{
Census data from 2017 and estimates for 2014 and 2018, as provided
by the Korean Statistical Information Service
}
\keyword{datasets}
