% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_projected_seroprevalence.r
\name{sk_projected_seroprevalence}
\alias{sk_projected_seroprevalence}
\title{Groups seroprevalence levels and projects to 2018}
\usage{
sk_projected_seroprevalence(contacts, equivocal = c("positive",
  "negative"))
}
\arguments{
\item{contacts}{a contact matrix}

\item{equivocal}{whether equivocal samples are to be interpreted as
"positive" (default) or "negative"}
}
\value{
table of seroprevalence by age group, original (2014) and shifted (2018)
}
\description{
Groups seroprevalence levels and projects to 2018
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
