% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{sk_seroprevalence}
\alias{sk_seroprevalence}
\title{Measles seroprevalence data for South Korea}
\format{A tibble with three columns: lower.age.limit (the lower bound
  in years of the age group), seropositive (the proportion seropositive) 
  and equivocal (the proportion equivocal).}
\source{
\url{https://doi.org/10.1016/j.vaccine.2017.06.058}
}
\usage{
sk_seroprevalence
}
\description{
A dataset containing measles seroprevalence in South Korea
measured as part of the KNHANES study in 2014.
The data are described in Kang et al. (2017) An increasing, 
potentially measles-susceptible population over time after 
vaccination in Korea. Vaccine 35(33): 4126-4132.
}
\keyword{datasets}
