% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.r
\name{split_20_29}
\alias{split_20_29}
\title{Sensitivity analysis splitting the 20-29 year age group}
\usage{
split_20_29(contacts)
}
\arguments{
\item{contacts}{a contact matrix}
}
\value{
contact matrix with 20-29 year olds split
}
\description{
Sensitivity analysis splitting the 20-29 year age group
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
