% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.r
\name{table1}
\alias{table1}
\title{Print table 1}
\usage{
table1(contacts, ...)
}
\arguments{
\item{contacts}{a contact matrix}

\item{...}{any further parameters to pass to \code{\link{sk_projected_seroprevalence}}}
}
\value{
wide table of seroprevalence by age group, original (2014) and shifted (2018)
}
\description{
Print table 1
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
